Changed StaxAppData to a static library. Modified output directory to Lib\x64\debug Finally I can link!

+++ StaxAppData/AppData.cpp
@@ -93,5 +93,5 @@ bool StaxPayments::AppData::RemoveDataItem(const StaxData& toRemove)
 
 int StaxPayments::AppData::GetListCount() const
 {
-  return _staxList.size();
+  return (int)_staxList.size();
 }
